+++ provider/devices/common.go
@@ -18,9 +18,9 @@ import (
 	"strings"
 	"time"
 
+	"GADS/common/constants"
 	"GADS/common/models"
 	"GADS/provider/config"
-	"GADS/provider/constants"
 	"GADS/provider/db"
 	"GADS/provider/logger"
 	"GADS/provider/util"
